+++ src/main/java/kr/co/preq/domain/auth/controller/AuthController.java
@@ -3,7 +3,7 @@ package kr.co.preq.domain.auth.controller;
 import javax.validation.Valid;
 
 import kr.co.preq.domain.auth.dto.request.AuthRequestDto;
-import kr.co.preq.domain.board.dto.response.response.AuthResponseDto;
+import kr.co.preq.domain.auth.dto.response.AuthResponseDto;
 import kr.co.preq.domain.auth.dto.request.LogoutRequestDto;
 import kr.co.preq.domain.auth.dto.request.TokenRequestDto;
 import kr.co.preq.domain.auth.service.AuthService;
